Revert "perf: update detail drawer show"
This reverts commit 3361b00bd176c76e94bea7bb1f30fd635af672e3.
