+++ src/views/users/User/UserDetail/UserLoginACL.vue
@@ -1,6 +1,6 @@
 <template>
   <TwoCol>
-    <UserLoginACLTable />
+    <UserLoginACLTable :url="url" />
   </TwoCol>
 </template>
